separated game files into own folders for clarity
