+++ app/src/main/AndroidManifest.xml
@@ -141,7 +141,7 @@
         <activity
             android:name=".ui.WorkoutSchedulingScreen" />
        <activity
-            android:name=".ui.GameActivity"/>
+            android:name=".ui.GameUI.GameActivity"/>
         <receiver
             android:name=".receiver.AlarmReceiver"
             android:process=":remote" />
