+++ app/src/main/java/org/liberty/android/fantastischmemo/modules/FragmentComponents.java
@@ -10,7 +10,7 @@ import org.liberty.android.fantastischmemo.ui.FileBrowserFragment;
 import org.liberty.android.fantastischmemo.ui.GestureSelectionDialogFragment;
 import org.liberty.android.fantastischmemo.ui.GradeButtonsFragment;
 import org.liberty.android.fantastischmemo.ui.MiscTabFragment;
-import org.liberty.android.fantastischmemo.ui.MultiPlayerRegistrationFragment;
+import org.liberty.android.fantastischmemo.ui.GameUI.MultiPlayerRegistrationFragment;
 import org.liberty.android.fantastischmemo.ui.OpenActionsFragment;
 import org.liberty.android.fantastischmemo.ui.QuizLauncherDialogFragment;
 import org.liberty.android.fantastischmemo.ui.RecentListFragment;
