+++ app/src/main/java/org/liberty/android/fantastischmemo/ui/GameUI/GameJoinFragment.java
@@ -1,4 +1,4 @@
-package org.liberty.android.fantastischmemo.ui;
+package org.liberty.android.fantastischmemo.ui.GameUI;
 
 import android.app.Activity;
 import android.os.Bundle;
