--- app/src/main/java/org/liberty/android/fantastischmemo/ui/TagsActivity.java
@@ -1,1 +1,0 @@
-//tags logic should be placed here
