+++ app/src/main/java/org/liberty/android/fantastischmemo/utils/Connections/ServerSender.java
@@ -1,4 +1,4 @@
-package org.liberty.android.fantastischmemo.utils;
+package org.liberty.android.fantastischmemo.utils.Connections;
 
 import java.io.IOException;
 import java.io.ObjectOutputStream;
