+++ app/src/main/res/layout/game_room.xml
@@ -12,9 +12,8 @@
         android:id="@+id/view"
         android:layout_width="280dp"
         android:layout_height="470dp"
-        android:layout_alignParentBottom="true"
-        android:layout_alignParentStart="true"
-        android:layout_marginStart="14dp"
+        android:layout_alignParentTop="true"
+        android:layout_alignStart="@+id/menuText"
         android:alpha="0.3"
         android:tag="gameRoom" />
 
@@ -35,34 +34,19 @@
         android:id="@+id/gameList"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:layout_above="@+id/gameSettings"
-        android:layout_alignEnd="@+id/menuText"
-        android:layout_alignStart="@+id/menuText"
+        android:layout_alignParentStart="true"
         android:layout_below="@+id/menuText"
-        android:layout_marginBottom="10dp"
-        android:layout_marginTop="5dp"
+        android:layout_marginTop="77dp"
         android:alpha="0.5"
         android:background="#ffffff" />
 
 
     <Button
-        android:id="@+id/gameSettings"
-        android:layout_width="150dp"
-        android:layout_height="wrap_content"
-        android:layout_alignParentBottom="true"
-        android:layout_alignParentEnd="true"
-        android:background="@drawable/button_drawable"
-        android:text="Game Settings"
-        android:textAllCaps="true"
-        android:textAppearance="?android:attr/textAppearanceSmall" />
-
-    <Button
         android:id="@+id/playGame"
         android:layout_width="150dp"
         android:layout_height="wrap_content"
-        android:layout_alignBaseline="@+id/gameSettings"
-        android:layout_alignBottom="@+id/gameSettings"
-        android:layout_alignParentStart="true"
+        android:layout_alignParentBottom="true"
+        android:layout_centerHorizontal="true"
         android:background="@drawable/button_drawable"
         android:text="Play Game"
         android:textAllCaps="true"
